seperate start project code out
