+++ modules/server/src/index.js
@@ -1,4 +1,5 @@
 import fs from 'fs';
+import elasticsearch from 'elasticsearch';
 import { rainbow } from 'chalk-animation';
 import { Server } from 'http';
 import express from 'express';
@@ -9,17 +10,11 @@ import cors from 'cors';
 import projectsRoutes from './projects';
 import sockets from './sockets';
 import { getProjects } from './utils/projects';
-import { PORT } from './utils/config';
+import startProject from './startProject';
+import { PORT, ES_HOST, PROJECT_ID } from './utils/config';
 
-let main = async () => {
-  let app = express();
-  let http = Server(app);
-  let io = socketIO(http);
-
+let startCMS = async ({ io, app }) => {
   sockets({ io });
-
-  app.use(cors());
-  app.use(bodyParser.json({ limit: '50mb' }));
 
   app.use((req, res, next) => {
     let projects = getProjects();
@@ -28,8 +23,31 @@ let main = async () => {
   });
 
   projectsRoutes({ app, io });
-
-  http.listen(PORT, () => rainbow(`⚡️ Listening on port ${PORT} ⚡️`));
 };
 
-main();
+let startSingleProject = async ({ app, io }) => {
+  sockets({ io });
+
+  const projectApp = await startProject({
+    es: new elasticsearch.Client({ host: ES_HOST }),
+    io,
+    id: PROJECT_ID,
+  });
+
+  app.use('/', projectApp);
+};
+
+let app = express();
+let http = Server(app);
+let io = socketIO(http);
+
+app.use(cors());
+app.use(bodyParser.json({ limit: '50mb' }));
+
+if (PROJECT_ID && ES_HOST) {
+  startSingleProject({ io, app });
+} else {
+  startCMS({ io, app });
+}
+
+http.listen(PORT, () => rainbow(`⚡️ Listening on port ${PORT} ⚡️`));
